update CUDA_TRY macross to explicitly cast cuda status to int (#514)
This change is needed for compatibility with compile-time checks in fmt v10.1.1

+++ operators/velodyne_lidar/cpp/velodyne_lidar.cpp
@@ -33,7 +33,7 @@
                          __LINE__,                                                      \
                          __FILE__,                                                      \
                          cudaGetErrorString(cuda_status),                               \
-                         cuda_status);                                                  \
+                         static_cast<int>(cuda_status));                                \
       throw std::runtime_error("Unable to copy device to host");                        \
     }                                                                                   \
   }
